+++ dbt/include/databricks/macros/relations/table/create.sql
@@ -3,12 +3,13 @@
   {% set model_columns = model.get('columns', []) %}
   {% set existing_columns = adapter.get_columns_in_relation(intermediate_relation) %}
   {% set contract_config = config.get('contract') %}
-  {% if contract_config and contract_config.enforced %}
+  {% set contract_enforced = contract_config and contract_config.enforced %}
+  {% if contract_enforced %}
     {% set model_constraints = model.get('constraints', []) %}
   {% else %}
     {% set model_constraints = [] %}
   {% endif %}
-  {% set columns_and_constraints = adapter.parse_columns_and_constraints(existing_columns, model_columns, model_constraints) %}
+  {% set columns_and_constraints = adapter.parse_columns_and_constraints(existing_columns, model_columns, model_constraints, contract_enforced) %}
   {% set target_relation = relation.enrich(columns_and_constraints[1]) %}
   
   {% call statement('main') %}